Better checking if blk-mq is in use
Improve the checking of blk-mq's use with the directory "mq" which is
created by blk-mq itself when enabled.
Since blk-mq has full support for I/O schedulers we don't have to
use only "none" as scheduler anymore.
To retrieve the available schedulers we use the same method for
both blk and blk-mq.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>