Adapt the save function to different test type and code refactoring
With the debug option the user can run different type of test with
different type of worloads and schedulers and the current
implementation of the save function does not support it, displaying
the results with wrong format.
Adapt the save function to this dynamism so that only the correct
results are dispayed.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>