Fix fiojob file creation
When creating a fio jobfile and we use more threads,
for example with debug option, for the fio command,
the generated jobfile causes several fio's errors due to
the uncorrect mask number.
So fix this problem by pinning all the fio
thread > max_cpu_count to the last cpu available.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>